// migrations/app/20260830T2359_criar_restaurantes/migration.ts
#!/usr/bin/env -S node
import type { Contract as End } from '../../snapshots/c82e922ad87297e46d394cf462cfa482ccc40d9c24918934e37af2522c1a6b7d/contract';
import endContract from '../../snapshots/c82e922ad87297e46d394cf462cfa482ccc40d9c24918934e37af2522c1a6b7d/contract.json' with { type: 'json' };
import { Migration, MigrationCLI, col, primaryKey } from '@prisma/orm-postgres/migration';

export default class M extends Migration<never, End> {
  override readonly endContractJson = endContract;

  override get operations() {
    return [
      this.createSchema({ schema: 'public' }),
      this.createTable({
        schema: 'public',
        table: 'restaurante',
        columns: [
          col('category', 'text', { notNull: true, codecRef: { codecId: 'pg/text@1' } }),
          col('id', 'SERIAL', { notNull: true, codecRef: { codecId: 'pg/int4@1' } }),
          col('name', 'text', { notNull: true, codecRef: { codecId: 'pg/text@1' } }),
          col('rating', 'float8', { notNull: true, codecRef: { codecId: 'pg/float8@1' } }),
        ],
        constraints: [primaryKey(['id'])],
      }),
    ];
  }
}

MigrationCLI.run(import.meta.url, M);
